Move matrix to own module
